# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Property {
  id: number;
  propertyId: string;
  titleEn: string;
  titleAr: string;
  propertyType: string;
  operation: string;
  neighborhood: string;
  /** @nullable */
  compound?: string | null;
  price: number;
  status: string;
  featured: boolean;
  area: number;
  bedrooms: number;
  bathrooms: number;
  /** @nullable */
  finishing?: string | null;
  /** @nullable */
  floor?: string | null;
  elevator?: boolean;
  imageUrl: string;
  imageUrls?: string[];
  /** @nullable */
  descriptionEn?: string | null;
  /** @nullable */
  descriptionAr?: string | null;
  createdAt?: string;
}

export interface PropertyInput {
  titleEn: string;
  titleAr: string;
  propertyType: string;
  operation: string;
  neighborhood: string;
  compound?: string;
  price: number;
  status?: string;
  featured?: boolean;
  area: number;
  bedrooms: number;
  bathrooms: number;
  finishing?: string;
  floor?: string;
  elevator?: boolean;
  imageUrl?: string;
  imageUrls?: string[];
  descriptionEn?: string;
  descriptionAr?: string;
}

export type PropertyUpdate = PropertyInput;

export interface Lead {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  propertyId?: number | null;
  /** @nullable */
  propertyTitle?: string | null;
  /** @nullable */
  message?: string | null;
  source: string;
  status: string;
  createdAt: string;
}

export interface LeadInput {
  name: string;
  phone: string;
  email?: string;
  propertyId?: number;
  message?: string;
  source: string;
}

export interface DashboardSummary {
  totalProperties: number;
  availableProperties: number;
  soldProperties: number;
  rentedProperties: number;
  featuredProperties: number;
  newLeads: number;
  recentProperties: Property[];
  recentLeads: Lead[];
}

export type ListPropertiesParams = {
operation?: string;
neighborhood?: string;
propertyType?: string;
minPrice?: number;
maxPrice?: number;
bedrooms?: number;
featured?: boolean;
search?: string;
};