# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Lead } from './lead';
import type { Property } from './property';

export interface DashboardSummary {
  totalProperties: number;
  availableProperties: number;
  soldProperties: number;
  rentedProperties: number;
  featuredProperties: number;
  newLeads: number;
  recentProperties: Property[];
  recentLeads: Lead[];
}