# lib/api-zod/src/generated/types/lead.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Lead {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  propertyId?: number | null;
  /** @nullable */
  propertyTitle?: string | null;
  /** @nullable */
  message?: string | null;
  source: string;
  status: string;
  createdAt: Date;
}