# lib/api-zod/src/generated/types/leadInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LeadInput {
  name: string;
  phone: string;
  email?: string;
  propertyId?: number;
  message?: string;
  source: string;
}