# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Property {
  id: number;
  propertyId: string;
  titleEn: string;
  titleAr: string;
  propertyType: string;
  operation: string;
  neighborhood: string;
  /** @nullable */
  compound?: string | null;
  price: number;
  status: string;
  featured: boolean;
  area: number;
  bedrooms: number;
  bathrooms: number;
  /** @nullable */
  finishing?: string | null;
  /** @nullable */
  floor?: string | null;
  elevator?: boolean;
  imageUrl: string;
  imageUrls?: string[];
  /** @nullable */
  descriptionEn?: string | null;
  /** @nullable */
  descriptionAr?: string | null;
  createdAt?: Date;
}