# lib/api-zod/src/generated/types/propertyInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PropertyInput {
  titleEn: string;
  titleAr: string;
  propertyType: string;
  operation: string;
  neighborhood: string;
  compound?: string;
  price: number;
  status?: string;
  featured?: boolean;
  area: number;
  bedrooms: number;
  bathrooms: number;
  finishing?: string;
  floor?: string;
  elevator?: boolean;
  imageUrl?: string;
  imageUrls?: string[];
  descriptionEn?: string;
  descriptionAr?: string;
}