# lib/api-zod/src/generated/types/propertyUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyInput } from './propertyInput';

export type PropertyUpdate = PropertyInput;